citigo-charge-monitor-worker/requirements.txt
pydantic==1.9.*
pyhumps==3.5.*
requests==2.27.*
skodaconnect==1.*